+++ counterparty-core/counterpartycore/lib/cli/server.py
@@ -26,7 +26,8 @@ from counterpartycore.lib import (
 from counterpartycore.lib.api import apiserver as api_v2
 from counterpartycore.lib.api import apiv1, dbbuilder
 from counterpartycore.lib.cli import bootstrap, log
-from counterpartycore.lib.ledger.currentstate import BackendHeight, CurrentState
+from counterpartycore.lib.ledger.backendheight import BackendHeight
+from counterpartycore.lib.ledger.currentstate import CurrentState
 from counterpartycore.lib.parser import blocks, check, follow
 from counterpartycore.lib.utils import database, helpers
